+++ packages/twenty-front/src/modules/workflow/components/WorkflowFormFieldInput.tsx
@@ -1,4 +1,4 @@
-import { FormFieldInput } from '@/object-record/record-field/form-types/components/FormFieldInput';
+import { FormFieldInput } from '@/object-record/record-field/form-types/components/FormFieldInputBase';
 import { SortOrFilterChip } from '@/views/components/SortOrFilterChip';
 import SearchVariablesDropdown from '@/workflow/search-variables/components/SearchVariablesDropdown';
 import { extractVariableLabel } from '@/workflow/search-variables/utils/extractVariableLabel';
